+++ client/src/Components/UserProgressInfo.tsx
@@ -10,12 +10,14 @@ import {
   withStyles,
 } from '@material-ui/core';
 import EditOutlinedIcon from '@material-ui/icons/EditOutlined';
+import Skeleton from '@material-ui/lab/Skeleton';
 import moment from 'moment';
 import React, { useState } from 'react';
 import { UserProgressForm } from '../Forms/UserProgressForm';
 import { useMySeriesProgressQuery } from '../gql/queries';
 import { ActionType } from '../utils/constants';
 import { renderWatchStatus } from '../utils/enumRender';
+import { UserProgressInfoSkeleton } from './Skeletons/UserProgressInfoSkeleton';
 
 type Props = {
   seriesId: string;
@@ -87,213 +89,223 @@ export const UserProgressInfo = (props: Props) => {
 
   return (
     <div>
-      <Paper elevation={3} className={classes.paper}>
-        <Grid container spacing={3}>
-          <Grid item xs={12} className={classes.tableHeader}>
-            <Grid container spacing={3}>
-              <Grid item xs>
-                <Typography variant="h5">Watch Progress</Typography>
+      {loading && !progress?.mySeriesProgress ? (
+        <UserProgressInfoSkeleton />
+      ) : (
+        <Paper elevation={3} className={classes.paper}>
+          <Grid container spacing={3}>
+            <Grid item xs={12} className={classes.tableHeader}>
+              <Grid container spacing={3}>
+                <Grid item xs>
+                  <Typography variant="h5">Watch Progress</Typography>
+                </Grid>
+                {!progress?.mySeriesProgress ? (
+                  <Grid item>
+                    <Button
+                      color="primary"
+                      variant="contained"
+                      startIcon={<EditOutlinedIcon />}
+                      onClick={() => {
+                        setFormActionType(ActionType.CREATE);
+                        setShowForm(true);
+                      }}
+                    >
+                      Update
+                    </Button>
+                  </Grid>
+                ) : (
+                  <Grid item>
+                    <Button
+                      color="primary"
+                      variant="contained"
+                      startIcon={<EditOutlinedIcon />}
+                      onClick={() => {
+                        setFormActionType(ActionType.UPDATE);
+                        setShowForm(true);
+                      }}
+                    >
+                      Update
+                    </Button>
+                  </Grid>
+                )}
               </Grid>
-              {!loading && !progress?.mySeriesProgress?.id ? (
-                <Grid item>
-                  <Button
-                    color="primary"
-                    variant="contained"
-                    startIcon={<EditOutlinedIcon />}
-                    onClick={() => {
-                      setFormActionType(ActionType.CREATE);
-                      setShowForm(true);
-                    }}
-                  >
-                    Update
-                  </Button>
-                </Grid>
-              ) : (
-                <Grid item>
-                  <Button
-                    color="primary"
-                    variant="contained"
-                    startIcon={<EditOutlinedIcon />}
-                    onClick={() => {
-                      setFormActionType(ActionType.UPDATE);
-                      setShowForm(true);
-                    }}
-                  >
-                    Update
-                  </Button>
-                </Grid>
-              )}
             </Grid>
+            {loading && !progress?.mySeriesProgress ? (
+              <Grid item xs={12}>
+                <Skeleton key={`progress-skeleton-1`} />
+                <Skeleton key={`progress-skeleton-2`} width={'85%'} />
+                <Skeleton key={`progress-skeleton-3`} width={'50%'} />
+              </Grid>
+            ) : !loading && !progress?.mySeriesProgress ? (
+              <Grid item xs={12}>
+                <Typography variant="overline" style={{ fontSize: '1em' }}>
+                  No existing watch progress record found
+                </Typography>
+              </Grid>
+            ) : (
+              <Grid item xs={12}>
+                <Grid container spacing={3} className={classes.tableContent}>
+                  <Grid item xs={2}>
+                    <Typography>Watch Status</Typography>
+                  </Grid>
+                  <Grid item xs={2}>
+                    <Typography>
+                      {progress?.mySeriesProgress?.status &&
+                        renderWatchStatus(progress?.mySeriesProgress?.status)}
+                    </Typography>
+                  </Grid>
+                  <Grid item xs={2}>
+                    <Typography>Episodes</Typography>
+                  </Grid>
+                  <Grid item xs={2}>
+                    <Typography>
+                      {progress?.mySeriesProgress?.completed || '0'} /{' '}
+                      {progress?.mySeriesProgress?.series?.episodeCount}
+                    </Typography>
+                  </Grid>
+                  <Grid item xs={2}>
+                    <Typography>Last Updated</Typography>
+                  </Grid>
+                  <Grid item xs={2}>
+                    <Typography>
+                      {moment(
+                        progress?.mySeriesProgress?.updatedAt ||
+                          progress?.mySeriesProgress?.createdAt
+                      ).format('HH:mm DD/MM/YYYY')}
+                    </Typography>
+                  </Grid>
+                  <Grid item xs={2}>
+                    <Typography>Overall</Typography>
+                  </Grid>
+                  <Grid item xs={9}>
+                    {progress?.mySeriesProgress?.overall ? (
+                      <BorderLinearProgress
+                        variant="determinate"
+                        value={progress?.mySeriesProgress?.overall}
+                      />
+                    ) : (
+                      <BorderLinearProgress variant="determinate" value={0} />
+                    )}
+                  </Grid>
+                  <Grid item xs={1}>
+                    <Typography>
+                      {progress?.mySeriesProgress?.overall || 0} / 100
+                    </Typography>
+                  </Grid>
+                  <Grid item xs={2}>
+                    <Typography>Story</Typography>
+                  </Grid>
+                  <Grid item xs={3}>
+                    {progress?.mySeriesProgress?.story ? (
+                      <BorderLinearProgress
+                        variant="determinate"
+                        value={progress?.mySeriesProgress?.story}
+                      />
+                    ) : (
+                      <BorderLinearProgress variant="determinate" value={0} />
+                    )}
+                  </Grid>
+                  <Grid item xs={1}>
+                    <Typography>
+                      {progress?.mySeriesProgress?.story || 0} / 100
+                    </Typography>
+                  </Grid>
+                  <Grid item xs={2}>
+                    <Typography>Execution</Typography>
+                  </Grid>
+                  <Grid item xs={3}>
+                    {progress?.mySeriesProgress?.execution ? (
+                      <BorderLinearProgress
+                        variant="determinate"
+                        value={progress?.mySeriesProgress?.execution}
+                      />
+                    ) : (
+                      <BorderLinearProgress variant="determinate" value={0} />
+                    )}
+                  </Grid>
+                  <Grid item xs={1}>
+                    <Typography>
+                      {progress?.mySeriesProgress?.execution || 0} / 100
+                    </Typography>
+                  </Grid>
+                  <Grid item xs={2}>
+                    <Typography>Appeal</Typography>
+                  </Grid>
+                  <Grid item xs={3}>
+                    {progress?.mySeriesProgress?.appeal ? (
+                      <BorderLinearProgress
+                        variant="determinate"
+                        value={progress?.mySeriesProgress?.appeal}
+                      />
+                    ) : (
+                      <BorderLinearProgress variant="determinate" value={0} />
+                    )}
+                  </Grid>
+                  <Grid item xs={1}>
+                    <Typography>
+                      {progress?.mySeriesProgress?.appeal || 0} / 100
+                    </Typography>
+                  </Grid>
+                  <Grid item xs={2}>
+                    <Typography>Character</Typography>
+                  </Grid>
+                  <Grid item xs={3}>
+                    {progress?.mySeriesProgress?.character ? (
+                      <BorderLinearProgress
+                        variant="determinate"
+                        value={progress?.mySeriesProgress?.character}
+                      />
+                    ) : (
+                      <BorderLinearProgress variant="determinate" value={0} />
+                    )}
+                  </Grid>
+                  <Grid item xs={1}>
+                    <Typography>
+                      {progress?.mySeriesProgress?.character || 0} / 100
+                    </Typography>
+                  </Grid>
+                  <Grid item xs={2}>
+                    <Typography>Art</Typography>
+                  </Grid>
+                  <Grid item xs={3}>
+                    {progress?.mySeriesProgress?.art ? (
+                      <BorderLinearProgress
+                        variant="determinate"
+                        value={progress?.mySeriesProgress?.art}
+                      />
+                    ) : (
+                      <BorderLinearProgress variant="determinate" value={0} />
+                    )}
+                  </Grid>
+                  <Grid item xs={1}>
+                    <Typography>
+                      {progress?.mySeriesProgress?.art || 0} / 100
+                    </Typography>
+                  </Grid>
+                  <Grid item xs={2}>
+                    <Typography>Music</Typography>
+                  </Grid>
+                  <Grid item xs={3}>
+                    {progress?.mySeriesProgress?.sound ? (
+                      <BorderLinearProgress
+                        variant="determinate"
+                        value={progress?.mySeriesProgress?.sound}
+                      />
+                    ) : (
+                      <BorderLinearProgress variant="determinate" value={0} />
+                    )}
+                  </Grid>
+                  <Grid item xs={1}>
+                    <Typography>
+                      {progress?.mySeriesProgress?.sound || 0} / 100
+                    </Typography>
+                  </Grid>
+                </Grid>
+              </Grid>
+            )}
           </Grid>
-          {!loading && !progress?.mySeriesProgress?.id ? (
-            <Grid item xs={12}>
-              <Typography variant="overline" style={{ fontSize: '1em' }}>
-                No existing watch progress record found
-              </Typography>
-            </Grid>
-          ) : (
-            <Grid item xs={12}>
-              <Grid container spacing={3} className={classes.tableContent}>
-                <Grid item xs={2}>
-                  <Typography>Watch Status</Typography>
-                </Grid>
-                <Grid item xs={2}>
-                  <Typography>
-                    {progress?.mySeriesProgress?.status &&
-                      renderWatchStatus(progress?.mySeriesProgress?.status)}
-                  </Typography>
-                </Grid>
-                <Grid item xs={2}>
-                  <Typography>Episodes</Typography>
-                </Grid>
-                <Grid item xs={2}>
-                  <Typography>
-                    {progress?.mySeriesProgress?.completed || '0'} /{' '}
-                    {progress?.mySeriesProgress?.series?.episodeCount}
-                  </Typography>
-                </Grid>
-                <Grid item xs={2}>
-                  <Typography>Last Updated</Typography>
-                </Grid>
-                <Grid item xs={2}>
-                  <Typography>
-                    {moment(
-                      progress?.mySeriesProgress?.updatedAt ||
-                        progress?.mySeriesProgress?.createdAt
-                    ).format('HH:mm DD/MM/YYYY')}
-                  </Typography>
-                </Grid>
-                <Grid item xs={2}>
-                  <Typography>Overall</Typography>
-                </Grid>
-                <Grid item xs={9}>
-                  {progress?.mySeriesProgress?.overall ? (
-                    <BorderLinearProgress
-                      variant="determinate"
-                      value={progress?.mySeriesProgress?.overall}
-                    />
-                  ) : (
-                    <BorderLinearProgress variant="determinate" value={0} />
-                  )}
-                </Grid>
-                <Grid item xs={1}>
-                  <Typography>
-                    {progress?.mySeriesProgress?.overall || 0} / 100
-                  </Typography>
-                </Grid>
-                <Grid item xs={2}>
-                  <Typography>Story</Typography>
-                </Grid>
-                <Grid item xs={3}>
-                  {progress?.mySeriesProgress?.story ? (
-                    <BorderLinearProgress
-                      variant="determinate"
-                      value={progress?.mySeriesProgress?.story}
-                    />
-                  ) : (
-                    <BorderLinearProgress variant="determinate" value={0} />
-                  )}
-                </Grid>
-                <Grid item xs={1}>
-                  <Typography>
-                    {progress?.mySeriesProgress?.story || 0} / 100
-                  </Typography>
-                </Grid>
-                <Grid item xs={2}>
-                  <Typography>Execution</Typography>
-                </Grid>
-                <Grid item xs={3}>
-                  {progress?.mySeriesProgress?.execution ? (
-                    <BorderLinearProgress
-                      variant="determinate"
-                      value={progress?.mySeriesProgress?.execution}
-                    />
-                  ) : (
-                    <BorderLinearProgress variant="determinate" value={0} />
-                  )}
-                </Grid>
-                <Grid item xs={1}>
-                  <Typography>
-                    {progress?.mySeriesProgress?.execution || 0} / 100
-                  </Typography>
-                </Grid>
-                <Grid item xs={2}>
-                  <Typography>Appeal</Typography>
-                </Grid>
-                <Grid item xs={3}>
-                  {progress?.mySeriesProgress?.appeal ? (
-                    <BorderLinearProgress
-                      variant="determinate"
-                      value={progress?.mySeriesProgress?.appeal}
-                    />
-                  ) : (
-                    <BorderLinearProgress variant="determinate" value={0} />
-                  )}
-                </Grid>
-                <Grid item xs={1}>
-                  <Typography>
-                    {progress?.mySeriesProgress?.appeal || 0} / 100
-                  </Typography>
-                </Grid>
-                <Grid item xs={2}>
-                  <Typography>Character</Typography>
-                </Grid>
-                <Grid item xs={3}>
-                  {progress?.mySeriesProgress?.character ? (
-                    <BorderLinearProgress
-                      variant="determinate"
-                      value={progress?.mySeriesProgress?.character}
-                    />
-                  ) : (
-                    <BorderLinearProgress variant="determinate" value={0} />
-                  )}
-                </Grid>
-                <Grid item xs={1}>
-                  <Typography>
-                    {progress?.mySeriesProgress?.character || 0} / 100
-                  </Typography>
-                </Grid>
-                <Grid item xs={2}>
-                  <Typography>Art</Typography>
-                </Grid>
-                <Grid item xs={3}>
-                  {progress?.mySeriesProgress?.art ? (
-                    <BorderLinearProgress
-                      variant="determinate"
-                      value={progress?.mySeriesProgress?.art}
-                    />
-                  ) : (
-                    <BorderLinearProgress variant="determinate" value={0} />
-                  )}
-                </Grid>
-                <Grid item xs={1}>
-                  <Typography>
-                    {progress?.mySeriesProgress?.art || 0} / 100
-                  </Typography>
-                </Grid>
-                <Grid item xs={2}>
-                  <Typography>Music</Typography>
-                </Grid>
-                <Grid item xs={3}>
-                  {progress?.mySeriesProgress?.sound ? (
-                    <BorderLinearProgress
-                      variant="determinate"
-                      value={progress?.mySeriesProgress?.sound}
-                    />
-                  ) : (
-                    <BorderLinearProgress variant="determinate" value={0} />
-                  )}
-                </Grid>
-                <Grid item xs={1}>
-                  <Typography>
-                    {progress?.mySeriesProgress?.sound || 0} / 100
-                  </Typography>
-                </Grid>
-              </Grid>
-            </Grid>
-          )}
-        </Grid>
-      </Paper>
+        </Paper>
+      )}
       {showForm && (
         <UserProgressForm
           open={showForm}
