+++ client/src/pages/EpisodePage/index.tsx
@@ -1,8 +1,8 @@
 import { Grid } from '@material-ui/core';
 import React from 'react';
 import { useParams } from 'react-router-dom';
-import { EpisodeInfoTable } from '../../Components/EpisodeInfoTable';
-import { FilesTable } from '../../Components/FilesTable';
+import { EpisodeInfo } from '../../Components/EpisodeInfo';
+import { FileList } from '../../Components/FileList';
 import { withAuth } from '../../HOC/withAuth';
 
 const EpisodePage = () => {
@@ -10,11 +10,11 @@ const EpisodePage = () => {
   return (
     <div>
       <Grid container spacing={3}>
-        <Grid item xs={12}>
-          <EpisodeInfoTable episodeId={episodeId} key={episodeId} />
+        <Grid item xs={12} key={`episode-page-1`}>
+          <EpisodeInfo episodeId={episodeId} key={`${episodeId}-info`} />
         </Grid>
-        <Grid item xs={12}>
-          <FilesTable episodeId={episodeId} />
+        <Grid item xs={12} key={`episode-page-2`}>
+          <FileList episodeId={episodeId} key={`${episodeId}-files`} />
         </Grid>
       </Grid>
     </div>
