+++ client/src/pages/SeriesPage/index.tsx
@@ -1,7 +1,7 @@
 import { Grid } from '@material-ui/core';
 import React from 'react';
 import { useParams } from 'react-router-dom';
-import { SeriesInfoTable } from '../../Components/SeriesInfoTable';
+import { SeriesInfo } from '../../Components/SeriesInfo';
 import { UserProgressInfo } from '../../Components/UserProgressInfo';
 import { withAuth } from '../../HOC/withAuth';
 import { EpisodesTable } from '../../Tables/EpisodesTable';
@@ -13,7 +13,7 @@ const SeriesPage = () => {
     <div>
       <Grid container spacing={3}>
         <Grid item xs={12}>
-          <SeriesInfoTable seriesId={seriesId} key={`${seriesId}-info`} />
+          <SeriesInfo seriesId={seriesId} key={`${seriesId}-info`} />
         </Grid>
         <Grid item xs={12}>
           <UserProgressInfo seriesId={seriesId} key={`${seriesId}-progress`} />
